requirements.txt
streamlit
openai
python-dotenv
sqlalchemy
psycopg2